Refactor: 컨벤션에 따른 폴더 위치 변경
Loading, ScrollToTop component를 common 하위로 이동
